pyproject.toml
[project]
name = "ensap-captcha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "marimo>=0.18.4",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
]
